plot: improve inline docs
